+++ feture/chat/antiCombo.js
@@ -1,0 +1,11 @@
+import * as helperFunction from "../../helperFunctions.js";
+import Settings from "../../settings";
+const config = Settings
+
+register('chat', (event) => {
+  var message = helperFunction.removeColors((ChatLib.getChatMessage(event)).toString());
+  if (!config.chatAntiCombo) return;
+  if (message.includes("Kill Combo")) {
+    cancel(event)
+  }
+})
