+++ feture/chat/antiTip.js
@@ -1,0 +1,10 @@
+import * as helperFunction from "../../helperFunctions.js";
+import Settings from "../../settings";
+const config = Settings
+
+register('chat', (event) => {
+  var message = helperFunction.removeColors((ChatLib.getChatMessage(event)).toString());
+  if (!config.chatAntiTip) return;
+  if (!message.includes("tipped")) return;
+  cancel(event)
+})
