+++ feture/chat/coopChatPing.js
@@ -1,0 +1,8 @@
+import * as helperFunction from "../../helperFunctions.js";
+import Settings from "../../settings";
+const config = Settings
+
+register("chat", () => {
+  if (!config.chatCoopPing) return;
+  World.playSound("random.orb", 2, 1);
+}).setCriteria("&r&bCo-op > ${*}");
