+++ feture/chat/cyberBully.js
@@ -1,0 +1,12 @@
+import * as helperFunction from "../../helperFunctions.js";
+import Settings from "../../settings";
+const config = Settings
+
+register('chat', (event) => {
+  var message = helperFunction.removeColors((ChatLib.getChatMessage(event)).toString());
+  if (!config.chatCyberBully) return;
+  if (message.startsWith(" ☠") && message.includes("was killed by")) {
+    ChatLib.command(`pc ${message}`);
+    ChatLib.command(`ac ${message}`);
+  }
+})
